+++ integration/spark/app/src/test/java/io/openlineage/spark/agent/column/ColumnLevelLineageHiveTest.java
@@ -16,7 +16,6 @@ import io.openlineage.spark.agent.util.LastQueryExecutionSparkEventListener;
 import io.openlineage.spark.api.OpenLineageContext;
 import io.openlineage.spark3.agent.lifecycle.plan.column.ColumnLevelLineageUtils;
 import java.util.Arrays;
-import java.util.Optional;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.hadoop.fs.FileSystem;
@@ -80,7 +79,7 @@ class ColumnLevelLineageHiveTest {
 
     context =
         OpenLineageContext.builder()
-            .sparkSession(Optional.of(spark))
+            .sparkSession(spark)
             .sparkContext(spark.sparkContext())
             .openLineage(new OpenLineage(Versions.OPEN_LINEAGE_PRODUCER_URI))
             .queryExecution(queryExecution)