+++ integration/spark/spark2/src/test/java/io/openlineage/spark2/agent/lifecycle/plan/CreateDataSourceTableAsSelectCommandVisitorTest.java
@@ -16,7 +16,6 @@ import io.openlineage.spark.agent.lifecycle.plan.CreateDataSourceTableAsSelectCo
 import io.openlineage.spark.api.OpenLineageContext;
 import java.net.URI;
 import java.util.List;
-import java.util.Optional;
 import org.apache.spark.SparkContext;
 import org.apache.spark.sql.SparkSession;
 import org.apache.spark.sql.catalyst.TableIdentifier$;
@@ -49,7 +48,7 @@ class CreateDataSourceTableAsSelectCommandVisitorTest {
     CreateDataSourceTableAsSelectCommandVisitor visitor =
         new CreateDataSourceTableAsSelectCommandVisitor(
             OpenLineageContext.builder()
-                .sparkSession(Optional.of(session))
+                .sparkSession(session)
                 .sparkContext(session.sparkContext())
                 .openLineage(new OpenLineage(Versions.OPEN_LINEAGE_PRODUCER_URI))
                 .build());