+++ integration/spark/spark3/src/test/java/io/openlineage/spark3/agent/lifecycle/plan/DropTableVisitorTest.java
@@ -37,7 +37,7 @@ class DropTableVisitorTest {
   private static final String TABLE = "table";
   OpenLineageContext openLineageContext =
       OpenLineageContext.builder()
-          .sparkSession(Optional.of(mock(SparkSession.class)))
+          .sparkSession(mock(SparkSession.class))
           .sparkContext(mock(SparkContext.class))
           .openLineage(new OpenLineage(Versions.OPEN_LINEAGE_PRODUCER_URI))
           .build();